UPSTREAM: mb/google/brya/var/vell: Remove Rcomp settings
This patch removes Rcomp settings. In MRC design, it checks if the
Rcomp settings from the board is 0 or null, if so, it uses the
recommended Rcomp values. Otherwise, it uses the Rcomp settings passed
from the UPD. From the change history of MRC, we're chasing a moving
target. This RCOMP setting in coreboot is an old setting while the
Rcomp settins in MRC are optimized settings. Moving forward, if there
is a new stepping, it might be changed again which increases the
maintenance effort in coreboot. IMHO, we should let MRC to set the
optimized RCOMP values for the design.

BUG=b:219378758
TEST=emerge-byra coreboot chromeos-bootimage and boots up with QS and
     PRQ CPUs. Checks with MRC log and ensure the RCOMP settings are
     filled properly by MRC.

(cherry picked from commit 9d0fc3f39650d5b3bc24507d8c3e053074bf370d)

Signed-off-by: Gaggery Tsai <gaggery.tsai@intel.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62201
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I8547e187b74f9b2cee57ddad2883d60c05d0b9fb
GitOrigin-RevId: 9d0fc3f39650d5b3bc24507d8c3e053074bf370d
Change-Id: I08c893e2fc08cca142307006b30f96462acc9c2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3500912
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>