UPSTREAM: mb/google/brya: Add PEG and initial Nvidia dGPU ASL support
Some brya variants will use a GN20 series Nvidia GPU, which requires
quite a bit of ACPI support code to be written for it. This patch
lands a decent bit of the initial code for it on the brya platform,

1) PEG RTD3 methods
2) DGPU power operations (RTD3 and GCOFF, NVJT _DSM and other Methods)
3) NVOP _DSM method

There will be more support to come later, this is all written to
specifications from the Nvidia Software Design Guide for GN20.

BUG=b:214581763
TEST=build patch train

including:
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62931
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Reviewed-by: Subrata Banik <subratabanik@google.com>
Original-Change-Id: Ifce1610210e9636e87dda4b55c8287334adfcc42
GitOrigin-RevId: c85253337934e3dd2aa5fc6b55371f9b48ef3b4c
Change-Id: I9de159dd3884e11d47510cc5e0e7ad74e5518a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3658750
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>