UPSTREAM: cpu/x86/smm: Add weak SoC init and exit methods
This change provides hooks for the SoC so it can perform any
initialization and cleanup in the SMM handler.

For example, if we have a UART enabled firmware with DEBUG_SMI, the UART
controller could have been powered off by the OS. In this case we need
to power on the UART when entering SMM, and then power it off before we
exit. If the OS had the UART enabled when entering SMM, we should
snapshot the UART register state, and restore it on exit. Otherwise we
risk clearing some interrupt enable bits.

BUG=b:221231786, b:217968734
TEST=Build test guybrush

(cherry picked from commit e6cd6caf31d1e22b56a94c941f50b47e77bb41cd)

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62500
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Original-Change-Id: I946619cd62a974a98c575a92943b43ea639fc329
GitOrigin-RevId: e6cd6caf31d1e22b56a94c941f50b47e77bb41cd
Change-Id: I8c4d5d70efcf4546d90a756cf1858b5db8407897
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3519131
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Rob Barnes <robbarnes@google.com>