UPSTREAM: Documentation/releases: Move upcoming release section to the top
The list of releases will grow and in the current state each release
moves the section for the upcoming release more down making it less
visible.

Thus, switch both sections, so that the documentation for the
upcoming release is at the top. Also, invert the order of the previous
releases, so that the latest is at the top.

(cherry picked from commit 2ff2da11c482d2a462aad1f38d50a0976f678d19)

Signed-off-by: Felix Singer <felixsinger@posteo.net>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62411
Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Change-Id: I69987e035f38ae3ca14dbf5c7644d5292106a978
GitOrigin-RevId: 2ff2da11c482d2a462aad1f38d50a0976f678d19
Change-Id: I6bb7c91b6fd8c4ccc6fed9c0892b4656797e5692
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3494766
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>