UPSTREAM: mb/amd/chausie/devicetree: set PSPP policy to DXIO_PSPP_DISABLED
Right now, the PSPP policy that controls if the PCIe lanes can be
dynamically downgraded to a lower speed to save some power needs to be
disabled in order for the link training to be successful. Once this
feature is working, the PSPP policy will be switched to balanced again.

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/62923
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Change-Id: I85a06f322c4ddff25c3a858e2b79c84b36c48932
GitOrigin-RevId: b9ee6f351b4552f024edee3f1e1d72a4a09ec45a
Change-Id: I4393dc444bbd0e513330e00bb4600f361b893de6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3538523
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>