UPSTREAM: cpu/intel/fit: Clear the FIT table when setting pointer
When rebuilding coreboot the empty fit table added to added to CBFS
stays the same so the build process sees no reason to update the file.
In the meantime ifittool did update that file for instance to add
microcode update entries. So each time coreboot is rebuilt the entries
are appended to the FIT table which runs out of space at some point.
One way to deal with this is to clear the fit table when setting the
pointer inside the bootblock.

error with a filled FIT table.

TESTED: Now running 'make' again on prodrive/hermes does not report an
Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63036
Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Reviewed-by: Sean Rhodes <sean@starlabs.systems>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Change-Id: Ia20a489dc90a4ae704e9ee6d532766899f83ffcc
GitOrigin-RevId: 422fdceaa89d82566afd6d20cb3519f8a8575c20
Change-Id: Ic4c496d5c3c2c7648271bf6d599135544c313f31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3565943
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Bhanu Prakash Maiya <bhanumaiya@google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>