UPSTREAM: soc/mediatek/mt8186: Fix pmif setting for low power mode
The current pmif register setting for low power mode is incorrect,
which is causing suspend failure. The issue of suspend failure is that
SRCLKENA0 will not be pulled down. EC will not be informed AP is
suspending now becuase of this. Therefore, add pmif_spmi_set_lp_mode()
to correct the setting.

This implementation is based on chapter 3.7 in MT8186 Functional
Specification.

BUG=b:215639203
TEST=test of suspend and resume pass.

Signed-off-by: Zhiyong Tao <zhiyong.tao@mediatek.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63089
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Yu-Ping Wu <yupingso@google.com>
Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Original-Change-Id: I2d02198f19f9cb052fba612c02404a6af1a10adb
GitOrigin-RevId: a7477706a096af540a5ee0ef04d0fca921fdab45
Change-Id: I433713b0aee00f9ecb6f5280be075199c22fc0bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3560355
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>