UPSTREAM: mb/google/brya/variants/baseboard/brask: set GPP_D0 to GPO
Currently, we control the GPP_D0 in the flash_fp_mcu in order to
program the component's firmware. If we set this pin to NC, then we
can't control the GPP_D0 output low/high and that make the system fails
to program the component's firmware. This patch sets the GPP_D0 to GPO
to fix it.

BUG=b:204679292
BRANCH=firmware-brya-14505.B
TEST=program the component's firmware

Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63090
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Original-Change-Id: I2f58c324f807a067dbe338f044a33dc9622ca469
GitOrigin-RevId: b0769db48ff394b1e71a139bc7ccd5b92ca13789
Change-Id: I339f3a9782f08a951870c7810ab0c52772fa8a77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3555615
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>