UPSTREAM: cpu/x86/smm_module_load: Rewrite setup_stub
This code was hard to read as it did too much and had a lot of state
to keep track of.

It also looks like the staggered entry points were first copied and
only later the parameters of the first stub were filled in. This
means that only the BSP stub is actually jumping to the permanent
smihandler. On the APs the stub would jump to wherever c_handler
happens to point to, which is likely 0. This effectively means that on
APs it's likely easy to have arbitrary code execution in SMM which is a
security problem.

(cherry picked from commit d7c371619a287a3a74e23fc3fcff4793a12deba6)

Original-Change-Id: I42ef9d6a30f3039f25e2cde975086a1365ca4182
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/63478
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin L Roth <gaumless@tutanota.com>
GitOrigin-RevId: d7c371619a287a3a74e23fc3fcff4793a12deba6
Change-Id: Ib22b9d360840c7fd1c7e8e8cadb8b93a80ebedd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3675285
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>