UPSTREAM: cpu/x86/64bit: Generate static page tables from an assembly file
This removes the need for a tool to generate simple identity pages.
Future patches will link this page table directly into the stages on
some platforms so having an assembly file makes a lot of sense.

This also optimizes the size of the page of each 4K page by placing
the PDPE_table below the PDE.

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63725
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Change-Id: Ia1e31b701a2584268c85d327bf139953213899e3
GitOrigin-RevId: 597b9e9d71a2dd728e4192fa9a083004c96cdfbc
Change-Id: Iacf01a7b2cd23fcb2dd1b72727174328dcd093f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3608313
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>