UPSTREAM: mb/purism/librem_mini: Rework front status LED to show all disk activity
The front status LED on the Librem Mini is driven by the SATALED# GPIO
line configured for native function, so only shows disk activity for
SATA drives, but not NVMe. To allow it to show disk activity for NVMe
drives as well, reconfigure the GPIO as GPIO-OUT (rather than native
function), and configure it via ACPI so that the linux gpio-leds
driver will attach and use it accordingly.

This has the added benefit of allowing the user to reconfigure the
LED as they see fit via sysfs.

LED blinks during periods of disk activity (tested via 'stress').

Test: boot Linux (PureOS) on Librem Mini v2 with NVMe drive, observe status
Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63749
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I34c2a5f3fd1038266f4514544abfc1020da6f85b
GitOrigin-RevId: f759a6257c0de47a579ec5fdfbd9d88f063669ec
Change-Id: I19637dd13d34ceb56c48e39abb35e3f43545635f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3604001
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>