UPSTREAM: soc/amd/common/block/psp/psp_gen2: rename cmd_response to buffer
The cmd_response field in the pspv2_mbox struct is the buffer used to
pass data to the PSP and back to the x86 side, so rename it to buffer.
This also aligns the code a bit more with the reference code. Also
rename the wr_mbox_cmd_resp function to wr_mbox_buffer_ptr.

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63962
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Change-Id: I22c8971b07b3dedcc2e6e50e93c98d69ec7379e8
GitOrigin-RevId: 81d0d89613c2eb30763c8cd04b36625b856eef38
Change-Id: I56773069b6314d4b724a374dfa4824ef1e9b12dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3622103
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>