UPSTREAM: soc/amd/common/block/psp/psp_gen2: factor out pspv2_mbox_command union
The pspv2_mbox struct contained an unnamed union that covered the 32
bits of the command register of the PSP v2 mailbox. Since the pspv2_mbox
struct is mainly used for hardware register accesses and the union part
is mostly used to access the different bits before/after writing/reading
the command register, split this functionality. For the register access
a command field is added to the pspv2_mbox struct instead of the unnamed
union and for accessing the separate bits of the command register a new
named union is added.

Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/63963
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Change-Id: If3f00b6fd73c3f749154b77b940e6d5aa385ec49
GitOrigin-RevId: 63e7b7064137f1854256f07a8f49b986ea944080
Change-Id: I8c76f579de81d518c92e5d3896eb9f1d507f5bcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/3622104
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>